NFC: Refactoring header includes.
This is reducing the dependence on compiler/ast.h as ground work
for switching to the new parser.
