+++ src/compiler/parser.h
@@ -7,6 +7,8 @@
 
 namespace verona::compiler
 {
+  struct File;
+
   std::unique_ptr<verona::compiler::File>
   parse(Context& context, std::string name, std::istream& input);
 }
