+++ src/compiler/recursive_visitor.h
@@ -104,105 +104,4 @@ namespace verona::compiler
       this->visit_expr(*expr.right, args...);
     }
   };
-
-  template<typename... Args>
-  class RecursiveTypeVisitor : public TypeVisitor<void, Args...>
-  {
-  protected:
-    void visit_entity_type(const EntityTypePtr& ty, Args... args) override
-    {
-      this->visit_types(ty->arguments, args...);
-    }
-
-    void visit_static_type(const StaticTypePtr& ty, Args... args) override
-    {
-      this->visit_types(ty->arguments, args...);
-    }
-
-    void visit_capability(const CapabilityTypePtr& ty, Args... args) override {}
-
-    void visit_union(const UnionTypePtr& ty, Args... args) override
-    {
-      this->visit_types(ty->elements, args...);
-    }
-
-    void
-    visit_intersection(const IntersectionTypePtr& ty, Args... args) override
-    {
-      this->visit_types(ty->elements, args...);
-    }
-
-    void visit_range_type(const RangeTypePtr& ty, Args... args) override
-    {
-      this->visit_type(ty->lower, args...);
-      this->visit_type(ty->upper, args...);
-    }
-
-    void visit_has_field_type(const HasFieldTypePtr& ty, Args... args) override
-    {
-      this->visit_type(ty->view, args...);
-      this->visit_type(ty->read_type, args...);
-      this->visit_type(ty->write_type, args...);
-    }
-
-    void visit_delayed_field_view_type(
-      const DelayedFieldViewTypePtr& ty, Args... args) override
-    {
-      this->visit_type(ty->type, args...);
-    }
-
-    void visit_fixpoint_type(const FixpointTypePtr& ty, Args... args) override
-    {
-      this->visit_type(ty->inner, args...);
-    }
-
-    void visit_entity_of_type(const EntityOfTypePtr& ty, Args... args) override
-    {
-      this->visit_type(ty->inner, args...);
-    }
-
-    void visit_string_type(const StringTypePtr& ty, Args... args) override {}
-    void visit_type_parameter(const TypeParameterPtr& ty, Args... args) override
-    {}
-    void visit_unit_type(const UnitTypePtr& ty, Args... args) override {}
-    void visit_is_entity_type(const IsEntityTypePtr& ty, Args... args) override
-    {}
-    void visit_infer(const InferTypePtr& ty, Args... args) override {}
-    void visit_fixpoint_variable_type(
-      const FixpointVariableTypePtr& ty, Args... args) override
-    {}
-
-    void visit_apply_region(const ApplyRegionTypePtr& ty, Args... args) override
-    {
-      this->visit_type(ty->type, args...);
-    }
-
-    void visit_viewpoint_type(const ViewpointTypePtr& ty, Args... args) override
-    {
-      for (const auto& v : ty->variables)
-      {
-        this->visit_type(v, args...);
-      }
-      this->visit_type(ty->right, args...);
-    }
-
-    void visit_variable_renaming_type(
-      const VariableRenamingTypePtr& ty, Args... args) override
-    {
-      this->visit_type(ty->type, args...);
-    }
-
-    void visit_path_compression_type(
-      const PathCompressionTypePtr& ty, Args... args) override
-    {
-      for (const auto& [v, replacement] : ty->compression)
-      {
-        this->visit_type(replacement, args...);
-      }
-      this->visit_type(ty->type, args...);
-    }
-
-    void visit_indirect_type(const IndirectTypePtr& ty, Args... args) override
-    {}
-  };
 }
