+++ src/compiler/substitution.h
@@ -2,7 +2,6 @@
 // SPDX-License-Identifier: MIT
 #pragma once
 
-#include "compiler/ast.h"
 #include "compiler/freevars.h"
 #include "compiler/mapper.h"
 #include "compiler/printing.h"
