+++ src/compiler/traits.h
@@ -3,6 +3,7 @@
 #pragma once
 
 #include <type_traits>
+#include <utility>
 
 namespace verona::compiler
 {
