+++ src/compiler/type.h
@@ -10,6 +10,7 @@
 #include <memory>
 #include <optional>
 #include <set>
+#include <string>
 #include <variant>
 
 namespace verona::compiler
